IMAP: Fetch all messages in folder; Implement a number of IMAP actions

+++ logic/storage/mail-sql.js
@@ -75,8 +75,7 @@ export default class MailSQLDatabase extends MailDatabase {
     if (toResult) {
       toID = toResult.id;
     }
-    await this._db.run(SQL`INSERT OR IGNORE INTO email (folder, msgID, firstFrom, firstTo, subject, dateSent, dateReceived) VALUES ((SELECT id FROM folder WHERE fullPath = ${folder.fullPath} AND accountID = ${folder.account.accountID} LIMIT 1), ${msg.msgID}, ${fromID}, ${toID}, ${msg.subject}, ${msg.date}, ${Date.now()})`);
-    // TODO to
+    await this._db.run(SQL`INSERT OR IGNORE INTO email (folder, msgID, parentMsgID, UID, firstFrom, firstTo, subject, dateSent, dateReceived) VALUES ((SELECT id FROM folder WHERE fullPath = ${folder.fullPath} AND accountID = ${folder.account.accountID} LIMIT 1), ${msg.msgID}, ${msg.parentMsgID}, ${msg.UID}, ${fromID}, ${toID}, ${msg.subject}, ${msg.date}, ${Date.now()})`);
     // TODO body
   }
 
@@ -89,7 +88,7 @@ export default class MailSQLDatabase extends MailDatabase {
   async deleteMessage(folder, msg) {
     assert(folder instanceof MsgFolder);
     assert(msg instanceof EMail);
-    await this._db.run(SQL`DELETE FROM email WHERE msgID = ${msg.msgID} AND folder = (SELECT id FROM folder WHERE fullPath = ${folder.fullPath} AND accountID = ${folder.account.accountID} LIMIT 1)`);
+    await this._db.run(SQL`DELETE FROM email WHERE msgID = ${msg.msgID} AND folder = (SELECT id AND UID = ${msg.UID} FROM folder WHERE fullPath = ${folder.fullPath} AND accountID = ${folder.account.accountID} LIMIT 1)`);
     // TODO body
   }
 
@@ -101,9 +100,10 @@ export default class MailSQLDatabase extends MailDatabase {
    */
   async listMessagesInFolder(folder) {
     assert(folder instanceof MsgFolder);
-    let results = await this._db.all(SQL`SELECT msgID, subject, dateSent, dateReceived, fromT.emailAddress as fromEmailAddress, fromT.name as fromName, toT.emailAddress as toEmailAddress, toT.name as toName FROM email LEFT JOIN person AS fromT ON firstFrom = fromT.id LEFT JOIN person AS toT ON firstTo = toT.id WHERE folder = (SELECT id FROM folder WHERE fullPath = ${folder.fullPath} AND accountID = ${folder.account.accountID} LIMIT 1)`);
+    let results = await this._db.all(SQL`SELECT UID, msgID, parentMsgID, subject, dateSent, dateReceived, fromT.emailAddress as fromEmailAddress, fromT.name as fromName, toT.emailAddress as toEmailAddress, toT.name as toName FROM email LEFT JOIN person AS fromT ON firstFrom = fromT.id LEFT JOIN person AS toT ON firstTo = toT.id WHERE folder = (SELECT id FROM folder WHERE fullPath = ${folder.fullPath} AND accountID = ${folder.account.accountID} LIMIT 1)`);
     return results.map(result => {
       let email = new EMail(folder);
+      email.UID = result.UID;
       email.msgID = result.msgID;
       email.subject = result.subject;
       email.date = new Date(result.dateSent);
